feat: move MountainBike to mountain_bike directory
